LayoutTests/webrtc/rtcconfiguration-icecandidatepoolsize.html

<!doctype html>
<html>
<body>

<script src="../resources/js-test-pre.js"></script>
<p>This test checks that RTCConfiguration.iceCandidatePoolSize is set.</p>
<div id="console"></div>

<script>
pc = new RTCPeerConnection({iceCandidatePoolSize: 10});
shouldBe('pc.getConfiguration().iceCandidatePoolSize', '10');

pc = new RTCPeerConnection({iceCandidatePoolSize: undefined});
shouldBe('pc.getConfiguration().iceCandidatePoolSize', '0');

pc = new RTCPeerConnection({iceCandidatePoolSize: null});
shouldBe('pc.getConfiguration().iceCandidatePoolSize', '0');

shouldThrow('new RTCPeerConnection({iceCandidatePoolSize: -1})', '"TypeError: Error creating RTCPeerConnection"');
shouldThrow('new RTCPeerConnection({iceCandidatePoolSize: 1000})', '"TypeError: Error creating RTCPeerConnection"');
shouldThrow('new RTCPeerConnection({iceCandidatePoolSize: "string"})', '"TypeError: Error creating RTCPeerConnection"');

</script>

<script src="../resources/js-test-post.js"></script>

</body>
</html>
